upgrade kitura to swift 4 (and last version)
